feat(api): move #createOrganization usecase to right context
