+++ api/tests/organizational-entities/unit/domain/validators/organization-creation-validator_test.js
@@ -1,6 +1,6 @@
-import * as organizationCreationValidator from '../../../../lib/domain/validators/organization-creation-validator.js';
-import { EntityValidationError } from '../../../../src/shared/domain/errors.js';
-import { catchErr, expect } from '../../../test-helper.js';
+import * as organizationCreationValidator from '../../../../../src/organizational-entities/domain/validators/organization-creation-validator.js';
+import { EntityValidationError } from '../../../../../src/shared/domain/errors.js';
+import { catchErr, expect } from '../../../../test-helper.js';
 
 const MISSING_VALUE = '';
